+++ Src/Private/Get-AbrWinFOClusterQuorum.ps1
@@ -1,0 +1,66 @@
+function Get-AbrWinFOClusterQuorum {
+    <#
+    .SYNOPSIS
+        Used by As Built Report to retrieve Microsoft FailOver Cluster Quorum
+    .DESCRIPTION
+        Documents the configuration of Microsoft Windows Server in Word/HTML/Text formats using PScribo.
+    .NOTES
+        Version:        0.5.0
+        Author:         Jonathan Colon
+        Twitter:        @jcolonfzenpr
+        Github:         rebelinux
+        Credits:        Iain Brighton (@iainbrighton) - PScribo module
+
+    .LINK
+        https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.Windows
+    #>
+
+    [CmdletBinding()]
+    param (
+    )
+
+    begin {
+        Write-PScriboMessage "FailOverCluster InfoLevel set at $($InfoLevel.FailOverCluster)."
+        Write-PscriboMessage "Collecting Host FailOver Cluster Quorum information."
+    }
+
+    process {
+        try {
+            $Settings = Invoke-Command -Session $TempPssSession { Get-ClusterQuorum -Cluster $using:Cluster | Select-Object -Property * } | Sort-Object -Property Name
+            if ($Settings) {
+                Section -Style Heading3 "Quorum" {
+                    $OutObj = @()
+                    foreach  ($Setting in $Settings) {
+                        try {
+                            $inObj = [ordered] @{
+                                'Cluster' = $Setting.Cluster
+                                'Quorum Resource' = $Setting.QuorumResource
+                                'Quorum Type' = $Setting.QuorumType
+                            }
+                            $OutObj += [pscustomobject]$inobj
+                        }
+                        catch {
+                            Write-PscriboMessage -IsWarning $_.Exception.Message
+                        }
+                    }
+
+                    $TableParams = @{
+                        Name = "Quorum - $($Cluster.toUpper().split(".")[0])"
+                        List = $false
+                        ColumnWidths = 33, 34, 33
+                    }
+                    if ($Report.ShowTableCaptions) {
+                        $TableParams['Caption'] = "- $($TableParams.Name)"
+                    }
+                    $OutObj | Table @TableParams
+                }
+            }
+        }
+        catch {
+            Write-PscriboMessage -IsWarning $_.Exception.Message
+        }
+    }
+
+    end {}
+
+}